revert changes; remove src/utils.js
