+++ Gruntfile.js
@@ -8,22 +8,18 @@ module.exports = function (grunt) {
         paths: {
             src: {
                 app_sms: [
-                    'src/utils.js',
                     'src/sms/app.js'
                 ],
                 app_ussd: [
-                    'src/utils.js',
                     'src/ussd/app.js'
                 ],
                 prd_sms: [
                     'src/sms/index.js',
-                    'src/utils.js',
                     '<%= paths.src.app_sms %>',
                     'src/sms/init.js'
                 ],
                 prd_ussd: [
                     'src/ussd/index.js',
-                    'src/utils.js',
                     '<%= paths.src.app_ussd %>',
                     'src/ussd/init.js'
                 ],
